WIP refactor + preparation to support extracting annotation with current annotation processor
